+++ CHANGES.md
@@ -94,6 +94,8 @@
     -   Make code block language detection handle unspecified
         or unknown code block languages.  ([GH-284][])
     -   Fix precedence of inline code over inline links.
+    -   Improve error reporting for `markdown` and `markdown-open`.
+        ([GH-291][])
 
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
   [gh-216]: https://github.com/jrblevin/markdown-mode/issues/216
@@ -129,6 +131,7 @@
   [gh-276]: https://github.com/jrblevin/markdown-mode/issues/276
   [gh-277]: https://github.com/jrblevin/markdown-mode/pull/277
   [gh-284]: https://github.com/jrblevin/markdown-mode/issues/284
+  [gh-291]: https://github.com/jrblevin/markdown-mode/issues/291
   [gh-296]: https://github.com/jrblevin/markdown-mode/issues/296
 
 # Markdown Mode 2.3